backend/requirements.txt
fastapi
uvicorn[standard]
motor
beanie
pydantic[email]
python-multipart
passlib[bcrypt]
python-jose[cryptography]
websockets
python-dotenv
opencv-python
numpy
razorpay
requests
email-validator
